+++ cmd/mount_windows_test.go
@@ -80,7 +80,7 @@ func (suite *mountTestSuite) SetupTest() {
 	options = mountOptions{}
 	err := log.SetDefaultLogger("silent", common.LogConfig{Level: common.ELogLevel.LOG_DEBUG()})
 	if err != nil {
-		panic("Unable to set silent logger as default.")
+		panic(fmt.Sprintf("Unable to set silent logger as default: %v", err))
 	}
 }