+++ component/block_cache/block_cache_linux_test.go
@@ -500,7 +500,7 @@ func TestBlockCacheTestSuite(t *testing.T) {
 	bcsuite := new(blockCacheTestSuite)
 	err := log.SetDefaultLogger("silent", common.LogConfig{Level: common.ELogLevel.LOG_DEBUG()})
 	if err != nil {
-		panic("Unable to set silent logger as default.")
+		panic(fmt.Sprintf("Unable to set silent logger as default: %v", err))
 	}
 
 	suite.Run(t, bcsuite)